Added restrictions to username and password
Added isAdmin variable to User. Added new constructor with default false boolean values. Reordered getters and setters. Added restrictions to username and password. Added email field in register. - we can change this later. I just thought it was confusing for username to be email. Also changed longName to username in User to be more consistent with rest of project.